+++ lib/command.tsx
@@ -1,12 +1,6 @@
 import * as React from 'react';
+import { CommandProps } from './types';
 
-export interface CommandProps {
-    children?: React.ReactNode,
-    name: string,
-    description?: string,
-    component?: React.ReactElement
-    flags?: any
-}
 
 export const Command = (props: CommandProps) => {
     if (props.component) {
