+++ lib/router.tsx
@@ -1,9 +1,10 @@
-import React, { createContext, FC } from 'react';
+import React, { createContext } from 'react';
 import yargs from 'yargs-parser';
+import { RouterProps } from './types'
 
 export const store = createContext({});
 
-export const Router: FC<{argv: string[]}> = ({ children,argv }) => {
+export const Router = ({ children, argv }: RouterProps) => {
     let args = yargs(argv.splice(2));
     let inputs = args._;
     // @ts-ignore
